refactor: move gene header to component
