+++ src/app/Scenes/Onboarding/OnboardingPersonalization2/OnboardingGene.tsx
@@ -2,24 +2,17 @@ import { useNavigation } from "@react-navigation/native"
 import { OnboardingGeneQuery } from "__generated__/OnboardingGeneQuery.graphql"
 import { InfiniteScrollArtworksGridContainer } from "app/Components/ArtworkGrids/InfiniteScrollArtworksGrid"
 import { FullScreenLoadingImage } from "app/Components/FullScreenLoadingImage"
-import { Button, Flex, FollowButton, Screen, Spacer, Text } from "palette"
+import { Button, Flex, Screen } from "palette"
 import { Suspense } from "react"
-import { ImageBackground, ImageSourcePropType } from "react-native"
-import { graphql, useLazyLoadQuery, useMutation } from "react-relay"
-import { AnimatedTooltip } from "./Components/AnimatedTooltip"
+import { graphql, useLazyLoadQuery } from "react-relay"
+import { GeneHeader, images } from "./Components/GeneHeader"
 import { useOnboardingContext } from "./Hooks/useOnboardingContext"
 
-type OnboardingGeneId = "artists-on-the-rise" | "trove" | "our-top-auction-lots"
+export type OnboardingGeneId = "artists-on-the-rise" | "trove" | "our-top-auction-lots"
 
 interface OnboardingGeneProps {
   id: OnboardingGeneId
   description: string
-}
-
-const images: Record<OnboardingGeneId, ImageSourcePropType> = {
-  "artists-on-the-rise": require("images/CohnMakeAMountain.webp"),
-  trove: require("images/HirstTheWonder.webp"),
-  "our-top-auction-lots": require("images/HirstTheWonder.webp"),
 }
 
 const OnboardingGene: React.FC<OnboardingGeneProps> = ({ id, description }) => {
@@ -29,19 +22,6 @@ const OnboardingGene: React.FC<OnboardingGeneProps> = ({ id, description }) => {
     id,
   })
 
-  const [commit, isInFlight] = useMutation(FollowGeneMutation)
-
-  const handleFollowGene = () => {
-    commit({
-      variables: {
-        input: {
-          id: gene?.id,
-          unfollow: gene?.isFollowed,
-        },
-      },
-    })
-  }
-
   return (
     <Screen>
       <Screen.Background>
@@ -50,29 +30,7 @@ const OnboardingGene: React.FC<OnboardingGeneProps> = ({ id, description }) => {
           itemComponentProps={{
             onPress: (artworkID) => navigate("ArtworkScreen", { artworkID }),
           }}
-          HeaderComponent={() => (
-            <Flex pb={2}>
-              <ImageBackground style={{ height: 270 }} resizeMode="cover" source={images[id]}>
-                <Flex pt={6} px={2}>
-                  <Text variant="xl" color="white100">
-                    {gene?.name}
-                  </Text>
-                  <Spacer mt={2} />
-                  <Text variant="sm" color="white100">
-                    {description}
-                  </Text>
-                  <Spacer mt={2} />
-                  <FollowButton
-                    isFollowed={!!gene?.isFollowed}
-                    onPress={handleFollowGene}
-                    loading={isInFlight}
-                    disabled={isInFlight}
-                  />
-                </Flex>
-              </ImageBackground>
-              <AnimatedTooltip />
-            </Flex>
-          )}
+          HeaderComponent={() => <GeneHeader geneID={id} description={description} gene={gene!} />}
           FooterComponent={() => (
             <Flex p={2}>
               <Button block onPress={onDone}>
@@ -95,6 +53,7 @@ export const OnboardingGeneScreen: React.FC<OnboardingGeneProps> = (props) => (
     fallback={
       <FullScreenLoadingImage
         imgSource={images[props.id]}
+        spacerHeight="80px"
         loadingText={"Great choice" + "\n" + "We’re finding a collection for you"}
       />
     }
@@ -106,9 +65,7 @@ export const OnboardingGeneScreen: React.FC<OnboardingGeneProps> = (props) => (
 const OnboardingGeneScreenQuery = graphql`
   query OnboardingGeneQuery($id: String!) {
     gene(id: $id) {
-      name
-      isFollowed
-      id
+      ...GeneHeaderFragment_Gene
       internalID
       artworks: filterArtworksConnection(
         first: 100
@@ -123,17 +80,6 @@ const OnboardingGeneScreenQuery = graphql`
         forSale: true
       ) {
         ...InfiniteScrollArtworksGrid_connection
-      }
-    }
-  }
-`
-
-const FollowGeneMutation = graphql`
-  mutation OnboardingGeneFollowMutation($input: FollowGeneInput!) {
-    followGene(input: $input) {
-      gene {
-        id
-        isFollowed
       }
     }
   }
